Remove MonoChainItem
